Fix Pool allocator to call connectAsync during acquire.
The Pool allocator as written doesn't call connectAsync, which cause the
inner object to not get created if the inner object isn't created during
the constructor (sqlite3 has this issue).

Fixing that - we might need to move off of using generic-pool altogether
though - it seems to have pre-allocation problems that we still need to
look into.